fix e2e for theme Logo replacement (#17907)
